% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFunctions.R
\name{BFbridge}
\alias{BFbridge}
\title{Distances per coordinates}
\usage{
BFbridge(ech, gamma, gammaref = rep(1, length(gamma)))
}
\arguments{
\item{ech}{posterior sample}
}
\value{
List of distances per coordinates
}
\description{
Distances per coordinates
}
