% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFunctions.R
\name{BFbridgeIS}
\alias{BFbridgeIS}
\title{Compute Bayes Factor from Importance Sampling}
\usage{
BFbridgeIS(ech, gamma, gammaref = rep(1, length(gamma)))
}
\arguments{
\item{ech}{posterior sample in the logit scale}

\item{gamma}{values given the alpha parameter of the prior distribution on the range parameters}

\item{gammaref}{values of the reference model, by default 1 for all range parameters}
}
\value{
Bayes factor between model with prior given in gamma and reference model (gamma=1)
}
\description{
Compute Bayes Factor from Importance Sampling
}
