% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFunctions.R
\name{computeProbActive}
\alias{computeProbActive}
\title{Compute the Probabilities that Input Variables are Active}
\usage{
computeProbActive(ech, valmax = 100)
}
\arguments{
\item{ech}{posterior sample in the logit scale}

\item{valmax}{value for alpha parameter for the inert prior distribution}
}
\value{
vector of activeness probabilities
}
\description{
Compute the Probabilities that Input Variables are Active
}
