% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MCMC}
\alias{MCMC}
\title{Metropolis sampler after a Metropolis within Gibbs Sampler}
\usage{
MCMC(
  niterMwG,
  niterMH,
  parwalk,
  parinit,
  Rexp,
  tdensD,
  alpha,
  parprior,
  adaptive,
  calibration
)
}
\arguments{
\item{niterMwG}{number of iterations for the Metropolis within Gibbs}

\item{niterMH}{number of iterations for the Metropolis}
}
\value{
Posterior sample in the transformed space
}
\description{
Metropolis sampler after a Metropolis within Gibbs Sampler
}
\details{
The MwG sampler is used to estimate a covariance matrix for the random walk in the upcoming Metropolis sampler.
}
