% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Metropolis}
\alias{Metropolis}
\title{Metropolis classic sampler with given covariance for RW}
\usage{
Metropolis(
  niter,
  covwalk,
  parinit,
  Rexp,
  tdensD,
  alpha,
  parprior,
  adaptive,
  calibration
)
}
\arguments{
\item{niter}{number of iterations}

\item{covwalk}{covariance matrix for random walk}
}
\value{
Posterior sample in the transformed space
}
\description{
Metropolis classic sampler with given covariance for RW
}
