% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MetropoliswGibbs}
\alias{MetropoliswGibbs}
\title{Metropolis within Gibbs Sampler}
\usage{
MetropoliswGibbs(
  niter,
  parwalk,
  parinit,
  Rexp,
  tdensD,
  alpha,
  parprior,
  adaptive,
  calibration
)
}
\arguments{
\item{niter}{number of iterations}
}
\value{
Posterior sample in the transformed space
}
\description{
Metropolis within Gibbs Sampler
}
