% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toySimulator.R
\name{sim3}
\alias{sim3}
\title{Non Linear Simulator with Calibration Parameter in 3 dimension}
\usage{
sim3(x, theta)
}
\arguments{
\item{x}{Matrix of field design in dimension n \eqn{\times} p (with p>3).
Only the 3 first rows of x are taken into account}

\item{theta}{Calibration parameter of dimension 3}
}
\value{
vector of responses
}
\description{
Non Linear Simulator with Calibration Parameter in 3 dimension
}
