% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toySimulator.R
\name{simGal}
\alias{simGal}
\title{Non Linear Simulator with Calibration Parameter in any Dimension}
\usage{
simGal(x, theta, beta = rep(1, ncol(x)))
}
\arguments{
\item{x}{Matrix of field design in dimension n \eqn{\times} p.}

\item{theta}{Calibration parameter of dimension 3}

\item{beta}{binary vector. Same size as \code{theta} indicating
with 1 which dimensions are active}
}
\value{
vector of responses corresponding to a linearized Sobol function
}
\description{
Non Linear Simulator with Calibration Parameter in any Dimension
}
