% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsFunctions.R
\name{tensordist}
\alias{tensordist}
\title{Distances per coordinates}
\usage{
tensordist(X1)
}
\arguments{
\item{X1}{design matrix.}
}
\value{
List of distances per coordinates
}
\description{
Distances per coordinates
}
